// server/graphql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ContractAddress: any;
  Cursor: any;
  DateTime: any;
  Enum: any;
  bool: any;
  felt252: any;
  u8: any;
  u32: any;
  u64: any;
  u128: any;
  usize: any;
};
// ...
export type Market = {
  __typename?: 'Market';
  cash?: Maybe<Scalars['u128']>;
  drug_id?: Maybe<Scalars['Enum']>;
  entity?: Maybe<World__Entity>;
  game_id?: Maybe<Scalars['u32']>;
  location_id?: Maybe<Scalars['Enum']>;
  quantity?: Maybe<Scalars['usize']>;
};

export type MarketConnection = {
  __typename?: 'MarketConnection';
  edges?: Maybe<Array<Maybe<MarketEdge>>>;
  pageInfo: World__PageInfo;
  totalCount: Scalars['Int'];
};

export type MarketEdge = {
  __typename?: 'MarketEdge';
  cursor?: Maybe<Scalars['Cursor']>;
  node?: Maybe<Market>;
};
// ...
export type ModelUnion = Drug | Encounter | Game | Item | Leaderboard | Market | Player | RyoMeta;
// ...
export type World__Entity = {
  __typename?: 'World__Entity';
  createdAt?: Maybe<Scalars['DateTime']>;
  eventId?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['ID']>;
  keys?: Maybe<Array<Maybe<Scalars['String']>>>;
  models?: Maybe<Array<Maybe<ModelUnion>>>;
  updatedAt?: Maybe<Scalars['DateTime']>;
};
// ...
export type BlindedMarket = {
  __typename?: 'BlindedMarket';
  cash?: Maybe<Scalars['u128']>;
  drug_id?: Maybe<Scalars['Enum']>;
  entity?: Maybe<World__Entity>;
  game_id?: Maybe<Scalars['u32']>;
  location_id?: Maybe<Scalars['felt252']>;
  quantity?: Maybe<Scalars['felt252']>;
};

export type BlindedMarketConnection = {
  __typename?: 'BlindedMarketConnection';
  edges?: Maybe<Array<Maybe<BlindedMarketEdge>>>;
  pageInfo: World__PageInfo;
  totalCount: Scalars['Int'];
};

export type BlindedMarketEdge = {
  __typename?: 'BlindedMarketEdge';
  cursor?: Maybe<Scalars['Cursor']>;
  node?: Maybe<BlindedMarket>;
};